+++ model_company_profile2.go
@@ -23,7 +23,7 @@ type CompanyProfile2 struct {
 	// IPO date.
 	Ipo string `json:"ipo,omitempty"`
 	// Market Capitalization.
-	MarketCapitalization int64 `json:"marketCapitalization,omitempty"`
+	MarketCapitalization float32 `json:"marketCapitalization,omitempty"`
 	// Number of oustanding shares.
 	ShareOutstanding float32 `json:"shareOutstanding,omitempty"`
 	// Logo image.
